refactor, fixed url encoded body, response
