+++ tests/test_format_response.py
@@ -7,11 +7,9 @@ from pygments.lexers import HttpLexer, JsonLexer, TextLexer
 from rich.syntax import Syntax
 
 from vedro_httpx import Response
-from vedro_httpx._render_response import (
-    format_response_body,
-    format_response_headers,
-    render_response, render_request,
-)
+from vedro_httpx._render_headers import format_response_headers
+from vedro_httpx._render_request import render_request
+from vedro_httpx._render_response import format_response_body, render_response
 
 
 def test_format_response_no_headers():
@@ -207,7 +205,7 @@ def test_render_response_with_patch_request_form_urlencoded():
             url='http://get_url.com',
             params={'test': 1},
             headers={'User-Agent': 'pytest'},
-            data={'id': 1}
+            data={'id': 1, 'name': 'TestName'}
         )
 
     with when:
@@ -224,12 +222,12 @@ def test_render_response_with_patch_request_form_urlencoded():
             "accept-encoding: gzip, deflate",
             "connection: keep-alive",
             "user-agent: pytest",
-            "content-length: 4",
+            "content-length: 18",
             "content-type: application/x-www-form-urlencoded",
         ])
         assert isinstance(headers_syntax.lexer, HttpLexer)
 
         assert isinstance(body_syntax, Syntax)
-        assert body_syntax.code == '{\n    "id": "1"\n}'
-        assert isinstance(body_syntax.lexer, JsonLexer)
+        assert body_syntax.code == 'id=1\n&name=TestName'
+        assert isinstance(body_syntax.lexer, TextLexer)
